pyproject.toml
[project]
name = "sales-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic-postgresql-enum>=1.8.0",
    "alembic[tz]>=1.17.2",
    "enum34>=1.1.10",
    "eralchemy2>=1.4.1",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pygraphviz>=1.14",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0.45",
    "typing>=3.10.0.0",
]
